+++ src/services/fusionService/__tests__/fusionService.testFixtures.ts
@@ -57,6 +57,8 @@ export function createFusionServiceTestContext(): FusionServiceTestContext {
         resetAccounts: false,
         resetForms: false,
         fusionOwnerIsGlobalReviewer: false,
+        fusionEnableManualReview: true,
+        fusionEnableAutoMerge: false,
         fusionReportOnAggregation: false,
         fusionFormAttributes: ['email', 'firstName', 'lastName'],
         baseurl: 'https://example.identitynow.com',